Fix some things.
