+++ server/src/inputs/__test__/flight.test.ts
@@ -1,26 +1,12 @@
+import {createMockDataContext} from "server/src/utils/createMockDataContext";
 import {flightInputs} from "../flight";
 
 describe("flight input", () => {
   it("should start a flight", async () => {
-    const mockDataContext = {
-      flight: null,
-      server: {
-        plugins: [
-          {
-            id: "Test Plugin",
-            name: "Test Plugin",
-            active: true,
-            aspects: {
-              ships: [
-                {
-                  name: "Test Template",
-                },
-              ],
-            },
-          },
-        ],
-      },
-    } as any;
+    const mockDataContext = createMockDataContext();
+    mockDataContext.database.flight = null;
+
+    expect(mockDataContext.flight).toBeNull();
 
     const flight = await flightInputs.flightStart(mockDataContext, {
       flightName: "Test Flight",
@@ -33,9 +19,10 @@ describe("flight input", () => {
       ],
     });
     expect(mockDataContext.flight).toBeDefined();
+    if (!mockDataContext.flight) throw new Error("No flight created");
     expect(mockDataContext.flight.name).toEqual("Test Flight");
     expect(mockDataContext.flight.ships.length).toEqual(1);
-    expect(mockDataContext.flight.ships[0].components.identity.name).toEqual(
+    expect(mockDataContext.flight.ships[0].components.identity?.name).toEqual(
       "Test Ship"
     );
   });
