+++ server/src/utils/inputAuth.ts
@@ -1,7 +1,7 @@
 import {DataContext} from "./DataContext";
 
 export default function inputAuth(context: DataContext) {
-  if (!context.client.isHost)
+  if (!context.isHost)
     throw new Error(
       "Unauthorized. You must be host to perform that operation."
     );
